UPDATES changed URL layout. This new layout can be reached on docs page
