+++ backend/destinations_routes_router.py
@@ -2,15 +2,17 @@
 original author: Dominik Cedro
 created: 2025-01-15
 license: none
-description: Main script
+description: Contains endpoint declarations for destinations and routes
 """
+from datetime import datetime
 from typing import List
 from fastapi import  Header, APIRouter
 from bson import ObjectId
 from fastapi import HTTPException
 from user_router import extract_user_id_from_token, get_user_by_id
-from models import  UserResponse,  DestinationModel, BaseDestinationModel, RouteModel
-from database import collection_users
+from models import UserResponse, DestinationModel, BaseDestinationModel, RouteModel, RoutePlan
+from database import collection_users, collection_route_plans
+
 destinations_routes_router = APIRouter()
 
 
@@ -93,3 +95,60 @@ async def update_destination(destination: BaseDestinationModel, authorization: s
     collection_users.update_one({"_id": ObjectId(user_id)}, {"$set": {"destinations": [dest.dict() for dest in user.destinations]}})
 
     return UserResponse(**user.dict())
+
+
+# route plans
+
+@destinations_routes_router.post("/create_route_plan", response_model=RoutePlan)
+async def create_route_plan(route_plan: RoutePlan, authorization: str = Header(...)):
+    """
+    Create a new route plan for a specific user.
+
+    Args:
+        route_plan (RoutePlan): The route plan data.
+        authorization (str): Bearer token from authorization header.
+
+    Returns:
+        RoutePlan: The inserted route plan document.
+
+    Raises:
+        HTTPException 401: If the user is not authenticated.
+        HTTPException 500: If the insertion fails.
+    """
+    token = authorization.split(" ")[1]
+    user_id = extract_user_id_from_token(token)
+    if not user_id:
+        raise HTTPException(status_code=401, detail="Invalid token or user not authenticated")
+
+    route_plan_data = route_plan.dict()
+    route_plan_data["creator_id"] = user_id
+    route_plan_data["date_created"] = datetime.utcnow()
+
+    result = collection_route_plans.insert_one(route_plan_data)
+    if result.inserted_id:
+        return RoutePlan(**route_plan_data)
+    else:
+        raise HTTPException(status_code=500, detail="Failed to create route plan")
+
+
+@destinations_routes_router.get("/get_my_route_plans", response_model=List[RoutePlan])
+async def get_my_route_plans(authorization: str = Header(...)):
+    """
+    Retrieve all route plans for the authenticated user.
+
+    Args:
+        authorization (str): Bearer token from authorization header.
+
+    Returns:
+        List[RoutePlan]: A list of route plans created by the user.
+
+    Raises:
+        HTTPException 401: If the user is not authenticated.
+    """
+    token = authorization.split(" ")[1]
+    user_id = extract_user_id_from_token(token)
+    if not user_id:
+        raise HTTPException(status_code=401, detail="Invalid token or user not authenticated")
+
+    route_plans = list(collection_route_plans.find({"creator_id": user_id}))
+    return [RoutePlan(**route_plan) for route_plan in route_plans]
